Moves field into its own file
